Services Model Getters

+++ app/Http/Controllers/Api/Backend/Provider/ProviderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api\Backend\Provider;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Provider\CreateProviderRequest;
 use App\Http\Requests\Provider\UpdateProviderRequest;
+use App\Http\Resources\ProviderResource;
 use App\Models\Provider;
 use App\Repositories\ProviderRepository;
 use App\Services\Auth\AuthService;
@@ -65,19 +66,16 @@ class ProviderController extends Controller
                 (int)$request->get('count', null)
             );
         } else {
-            $providers = $this->providerService->findUserPermittedProviders(
+            $providers = $this->providerService->findUserProviders(
+                $user,
                 $request->get('sort', "name"),
                 $request->get('order', "asc"),
-                (int)$request->get('count', null),
-                $request->user()
+                (int)$request->get('count', null)
             );
         }
         return $this->sendSuccessResponse(
             "success",
-            $this->serializerService->entityArrayToArray(
-                $providers,
-                ["list"]
-            )
+            ProviderResource::collection($providers)
         );
     }
 
@@ -88,19 +86,20 @@ class ProviderController extends Controller
     public function getProvider(Provider $provider, Request $request): \Illuminate\Http\JsonResponse
     {
         $this->setAccessControlUser($request->user());
-        if (!$request->user()->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_SUPERUSER)) && !$request->user(
-            )->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_ADMIN))) {
-            $this->accessControlService->checkPermissionsForEntity(
+        if (
+            !$this->accessControlService->checkPermissionsForEntity(
                 $provider,
                 [
                     PermissionService::PERMISSION_ADMIN,
                     PermissionService::PERMISSION_READ,
                 ],
-            );
+            )
+        ) {
+            return $this->sendErrorResponse("Access control: operation not permitted");
         }
         return $this->sendSuccessResponse(
             "success",
-            $this->serializerService->entityToArray($provider, ["single"])
+            new ProviderResource($provider)
         );
     }
 
@@ -111,6 +110,7 @@ class ProviderController extends Controller
     public function createProvider(CreateProviderRequest $request): \Illuminate\Http\JsonResponse
     {
         $createProvider = $this->providerService->createProvider(
+            $request->user(),
             $request->all()
         );
 
@@ -119,7 +119,7 @@ class ProviderController extends Controller
         }
         return $this->sendSuccessResponse(
             "Provider added",
-            $this->serializerService->entityToArray($createProvider, ['main'])
+            new ProviderResource($this->providerService->getProviderRepository()->getModel())
         );
     }
 
@@ -131,15 +131,16 @@ class ProviderController extends Controller
     public function updateProvider(Provider $provider, UpdateProviderRequest $request): \Illuminate\Http\JsonResponse
     {
         $this->setAccessControlUser($request->user());
-        if (!$request->user()->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_SUPERUSER)) && !$request->user(
-            )->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_ADMIN))) {
-            $this->accessControlService->checkPermissionsForEntity(
+        if (
+            !$this->accessControlService->checkPermissionsForEntity(
                 $provider,
                 [
                     PermissionService::PERMISSION_ADMIN,
                     PermissionService::PERMISSION_UPDATE,
                 ],
-            );
+            )
+        ) {
+            return $this->sendErrorResponse("Access control: operation not permitted");
         }
         $updateProvider = $this->providerService->updateProvider(
             $provider,
@@ -163,26 +164,26 @@ class ProviderController extends Controller
     public function deleteProvider(Provider $provider, Request $request): \Illuminate\Http\JsonResponse
     {
         $this->setAccessControlUser($request->user());
-        if (!$request->user()->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_SUPERUSER)) && !$request->user(
-            )->tokenCan(AuthService::getApiAbility(AuthService::ABILITY_ADMIN))) {
-            $this->accessControlService->checkPermissionsForEntity(
+
+        if (
+            !$this->accessControlService->checkPermissionsForEntity(
                 $provider,
                 [
                     PermissionService::PERMISSION_ADMIN,
                     PermissionService::PERMISSION_DELETE,
                 ],
-            );
+            )
+        ) {
+            return $this->sendErrorResponse("Access control: operation not permitted");
         }
         $delete = $this->providerService->deleteProvider($provider);
         if (!$delete) {
             return $this->sendErrorResponse(
-                "Error deleting provider",
-                $this->serializerService->entityToArray($delete, ['main'])
+                "Error deleting provider"
             );
         }
         return $this->sendSuccessResponse(
-            "Provider deleted.",
-            $this->serializerService->entityToArray($delete, ['main'])
+            "Provider deleted."
         );
     }
 }
